fix: six correctness, efficiency, and architecture issues
Correctness:
- pending_ocr now cleared on erasure — prevents stale OCR results from
  writing to ledger after an entity is erased (main.py newly_erased loop)
- INFERRING-stuck entities fixed: empty crop resets entity back to
  STABILIZING via Registry.reset_to_stabilizing() rather than leaving it
  permanently stuck awaiting an OCR result that was never submitted
- Remove vestigial ocr_confidence field from SemanticEntity — VLM
  backends never return confidence; field was always written as 1.0 and
  never read; drop confidence param from mark_active() accordingly

Runtime:
- BoardReconstructor.update() short-circuits cv2.distanceTransform when
  mask has no person pixels (mask.any() == False), replacing a 2M-pixel
  distance transform with a uniform in-place EMA — covers the common case
  where no lecturer is in frame
- HDBSCANGrouper: compute pairwise distance matrix in float64 natively
  (matching HDBSCAN's requirement), eliminating the float32→float64
  recast allocation in group()

Architecture:
- Flatten subprocess-within-subprocess: TextLineDetector is now
  synchronous — PaddleOCR runs directly inside the stage5-layout worker
  process which already provides async isolation to main. Eliminates the
  paddle-detect grandchild process, internal queues, one-frame result lag,
  and ~60 lines of queue/process management code.
  TextBlockDetector.load() now calls line_detector.load() explicitly.

+++ src/layout_service/hdbscan_grouper.py
@@ -25,7 +25,7 @@ class HDBSCANGrouper(TextLineGrouper):
         scale = np.maximum((h + h.T) / 2.0, 1e-5)
         norm_dx = dx / (self.horizontal_scale * scale)
         norm_dy = dy / scale
-        return np.sqrt(norm_dx**2 + norm_dy**2).astype(np.float32)
+        return np.sqrt(norm_dx**2 + norm_dy**2)
 
     def group(self, lines: list[TextLine]) -> list[Block]:
         if not lines:
@@ -42,7 +42,7 @@ class HDBSCANGrouper(TextLineGrouper):
             h = line.bbox[3] - line.bbox[1]
             centroids.append([cx, cy, h])
 
-        X = np.array(centroids, dtype=np.float32)
+        X = np.array(centroids, dtype=np.float64)
         dist_matrix = self._custom_pairwise_distance(X)
 
         clusterer = HDBSCAN(
@@ -50,7 +50,7 @@ class HDBSCANGrouper(TextLineGrouper):
             min_cluster_size=self.min_cluster_size,
             min_samples=1,
         )
-        labels = clusterer.fit_predict(dist_matrix.astype(np.float64))
+        labels = clusterer.fit_predict(dist_matrix)
 
         groups_dict = {}
         for idx, label in enumerate(labels):
